refactor: created bridges package and moved ros_bridge within it
