+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros_bridge/adapter_base.py
