+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros_bridge/data_ontology/battery_state.py
