+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros_bridge/injector.py
