+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros_bridge/ros_message.py
