+++ mosaico-sdk-py/src/mosaicolabs/bridges/ros_bridge/sequence_extractor.py
@@ -28,12 +28,12 @@ from rosbags.typesys import Stores, get_typestore
 from rosbags.typesys.store import Typestore
 
 from mosaicolabs import Message, MosaicoClient
+from mosaicolabs.bridges.ros_bridge.adapter_base import ROSAdapterBase
+from mosaicolabs.bridges.ros_bridge.loader import MosaicoLoader
+from mosaicolabs.bridges.ros_bridge.qos import get_qos_for_topic
+from mosaicolabs.bridges.ros_bridge.registry import ROSTypeRegistry
+from mosaicolabs.bridges.ros_bridge.ui import ProgressManager
 from mosaicolabs.logging_config import get_logger, setup_sdk_logging
-from mosaicolabs.ros_bridge.adapter_base import ROSAdapterBase
-from mosaicolabs.ros_bridge.loader import MosaicoLoader
-from mosaicolabs.ros_bridge.qos import get_qos_for_topic
-from mosaicolabs.ros_bridge.registry import ROSTypeRegistry
-from mosaicolabs.ros_bridge.ui import ProgressManager
 
 # Set the hierarchical logger
 logger = get_logger(__name__)
