+++ mosaico-sdk-py/src/testing/integration/ros_bridge/test_sequence_extractor.py
@@ -6,8 +6,8 @@ from rosbags.rosbag2 import StoragePlugin
 from rosbags.typesys import Stores, get_typestore
 
 from mosaicolabs import Pose, SessionLevelErrorPolicy
-from mosaicolabs.ros_bridge import MosaicoLoader, ROSBridge, ROSLoader
-from mosaicolabs.ros_bridge.sequence_extractor import (
+from mosaicolabs.bridges.ros_bridge import MosaicoLoader, ROSBridge, ROSLoader
+from mosaicolabs.bridges.ros_bridge.sequence_extractor import (
     ROSExtractorConfig,
     ROSSequenceExtractor,
 )
