+++ mosaico-sdk-py/src/testing/unit/ros_bridge/test_ros_schema_metadata.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mosaicolabs.ros_bridge.adapter_base import RosSchemaMetadata
+from mosaicolabs.bridges.ros_bridge.adapter_base import RosSchemaMetadata
 
 
 def test_to_dict_wraps_fields_under_reserved_key():
