+++ mosaico-sdk-py/src/testing/unit/ros_bridge/test_schema_metadata_from_injector.py
@@ -2,8 +2,8 @@ from pathlib import Path
 from types import SimpleNamespace
 from typing import Optional
 
-from mosaicolabs.ros_bridge.injector import RosbagInjector, ROSInjectionConfig
-from mosaicolabs.ros_bridge.ros_message import ROSMessage
+from mosaicolabs.bridges.ros_bridge.injector import RosbagInjector, ROSInjectionConfig
+from mosaicolabs.bridges.ros_bridge.ros_message import ROSMessage
 
 
 class _FakeSessionWriter:
